+++ src/xarray_regrid/__init__.py
@@ -1,11 +1,17 @@
 from xarray_regrid import methods
 from xarray_regrid.regrid import Regridder
 from xarray_regrid.utils import Grid, create_regridding_dataset
+from xarray_regrid.constants import GridType
+from xarray_regrid.core import BaseRegridder, RectilinearRegridder, CurvilinearRegridder
 
 __all__ = [
     "Grid",
     "Regridder",
+    "BaseRegridder",
+    "RectilinearRegridder",
+    "CurvilinearRegridder",
     "create_regridding_dataset",
+    "GridType",
     "methods",
 ]